Attempt to fix checkstyle error observed on Teamcity host
The Teamcity works with Git repository differently from a typical human
to optimize the checkout process. This results in the fact that list of branches
is different from one observed locally. For example, there is no origin/master present.
But there seems to be just master checked out.

I've changed the diff command to account for this.